feat(@galactiks/explorer): create multiple pages when locales with the same language exist
