+++ packages/explorer/package.json
@@ -39,6 +39,7 @@
     "@std-uritemplate/std-uritemplate": "0.0.49",
     "debug": "4.3.4",
     "esbuild": "0.19.11",
+    "lodash.groupby": "^4.6.0",
     "mdx-bundler": "10.0.1",
     "react": "18.2.0",
     "slugify": "1.6.6",
@@ -46,6 +47,7 @@
   },
   "devDependencies": {
     "@types/debug": "4.1.12",
+    "@types/lodash.groupby": "^4.6.9",
     "@types/react": "18.2.46",
     "schema-dts": "1.1.2"
   }
