+++ packages/explorer/src/core/content/hydrate/types.mts
@@ -1,0 +1,11 @@
+import type { GalactiksConfig } from '@galactiks/config';
+import type { Content, ContentlayerPerson, ContentlayerWebPageDocument, ContentlayerWebsite } from '../../types/index.mjs';
+
+export type ComputeDTO<T> = {
+  config: GalactiksConfig;
+  documents: T[];
+  websites: ContentlayerWebsite[];
+  people: ContentlayerPerson[];
+};
+
+export type ComputeHandler = (config: GalactiksConfig, websites: ContentlayerWebsite[], people: ContentlayerPerson[]) => (documents: ContentlayerWebPageDocument) => Promise<Content[]>
