+++ packages/explorer/src/core/content/index.mts
@@ -1,2 +1,2 @@
+export type * from '../types/index.mjs';
 export * from './repositories/index.mjs';
-export * from './types/index.mjs';
