+++ packages/explorer/src/core/content/metadata/breadcrumb.mts
@@ -1,6 +1,5 @@
-import type { ItemList } from '../types/index.mjs';
+import type { ContentlayerDocumentWithURL, ItemList } from '../../types/index.mjs';
 import { homeIdentifier, pageDepthLimit } from '../consts.mjs';
-import type { ContentlayerDocumentWithURL } from '../urls.mjs';
 import { documentByIdentifierAndLanguageSelector } from '../selectors.mjs';
 import { MaxDepthLimitReachedError } from '../../exceptions/index.mjs';
