+++ packages/explorer/src/core/content/metadata/headers.mts
@@ -1,4 +1,4 @@
-import type { Content, MetadataHeaders } from '../types/index.mjs';
+import type { Content, MetadataHeaders } from '../../types/index.mjs';
 
 export const getBasicHeaders = (entry: Content): MetadataHeaders => ({
   title: entry.name,
