+++ packages/explorer/src/core/content/metadata/schemas/organization.mts
@@ -1,5 +1,5 @@
 import type { Organization, WithContext } from 'schema-dts';
-import type { Organization as ContentOrganization } from '../../types/index.mjs';
+import type { Organization as ContentOrganization } from '../../../types/index.mjs';
 
 export const getOrganization = (
   document: ContentOrganization
