+++ packages/explorer/src/core/content/metadata/schemas/person.mts
@@ -1,5 +1,5 @@
 import type { Person, WithContext } from 'schema-dts';
-import type { Person as ContentPerson } from '../../types/index.mjs';
+import type { Person as ContentPerson } from '../../../types/index.mjs';
 
 export const getPerson = (document: ContentPerson): WithContext<Person> => ({
   '@context': 'https://schema.org',
