+++ packages/explorer/src/core/content/metadata/schemas/structured-data.mts
@@ -1,6 +1,6 @@
 import type { Thing, WithContext } from 'schema-dts';
 
-import type { Content } from '../../types/index.mjs';
+import type { Content } from '../../../types/index.mjs';
 import { getArticle } from './article.mjs';
 import { getBreadcrumb } from './breadcrumb.mjs';
