+++ packages/explorer/src/core/content/repositories/filters.mts
@@ -1,4 +1,4 @@
-import type { Content } from '../types/index.mjs';
+import type { Content } from '../../types/index.mjs';
 
 export type RepositoryFilters = {
   type?: Content['type'];
