+++ packages/explorer/src/core/types/render.mts
@@ -1,0 +1,6 @@
+import type { MDX } from '@galactiks/contentlayer';
+
+export type Render = () => { Content: React.FC };
+export type ContentlayerDocumentWithRender<T> = T & {
+  body: MDX & { render: Render };
+};
