+++ packages/explorer/tsconfig.json
@@ -7,6 +7,11 @@
     "target": "ES2021",
     "module": "ES2022",
     "declarationDir": "./dist",
-    "outDir": "./dist"
+    "outDir": "./dist",
+    "paths": {
+      "@/*": ["./*"],
+      "@content/*": ["./content/*"],
+      "@exceptions/*": ["./exceptions/*"],
+    }
   }
 }
